// ProductionProject/SeattleInventory/SeattleInventoryPortal/seattle-inventory-portal/src/app/model/inventoryorderitem.ts
export class inventoryorderitem{
    order_id: number ;
    inv_id : number ;
    inv_item_id   : number ;
    email :     string ;
    project:    string ;
    instdate:    Date ;
    destb:     string ;
    destf:     string ;
    room:    string ;
    qty     : number ;
    category:    string ;
    item_code :    string ;
    description:     string ;
    building:   string ;
    floor:    string ;
    mploc:     string ;
    imgBase64 :    string ;
    imgBase64Num: number;
    instdatestr: string;
}